Dispatch L2 distances using L2_options
The instance with rbf_fin_op caused some headache:

Because the handling of L2 expanded and unexpanded is unified in
``distance_impl_l2_with_options``, an instance of the CUTLASS distance
kernel for rbf_fin_op was instantiated. For some reason, CUTLASS did not
accept this as a valid argument and threw a very very big error message.
I could not get the rbf_fin_op in acceptable state for cutlass: I
included a default constructor, put const on every method, but to no
avail.

The current solution is to avoid CUTLASS when another final op is used
than the raft::identity_op.