Remove raft/matrix/matrix.cuh includes
The `raft/matrix/matrix.cuh` file has been marked as deprecated, and
produces a compile warning when included. However it was still being
referenced in a bunch of different spots within raft - making it
hard to avoid these warnings.

Remove the includes, in favour of either the newer API's or in certain
cases the detail API