Generate dataset of select_k times (#1497)
This adds an optional flag (`--select_k_dataset`) to the MATRIX_BENCH that will turn on generating a grid search of benchmarks for different select_k algorithms.  Since this adds about 100x as many benchmarks to run as previous (90k vs 900), this is opt-in only right now. This will be used to learn a heuristic function in #1455

This also integrates the faiss block select top-k algorithm into this benchmarking, so that we can compare how it performs against the other select_k algorithms

Authors:
  - Ben Frederickson (https://github.com/benfred)

Approvers:
  - Corey J. Nolet (https://github.com/cjnolet)

URL: https://github.com/rapidsai/raft/pull/1497