Learn heuristic to pick fastest select_k algorithm
This uses the select_k dataset from https://github.com/rapidsai/raft/pull/1497 to
learn a heuristic of the fastest select_k variant based off the rows/ cols/k
of the input. This heuristic is modelled as a DecisionTree, which is automatically
exported in C++ code that is compiled into RAFT. This lets us learn a function to
pick the fastest select_k method - which requires only a few if statements in C++
code, making it very cheap to evaluate.