[REVIEW] Fix padding assertion in sparse Gram evaluation (#1521)
A assertion could fail if the result kernel matrix is column major but has only a single row. In this case the old implementation identified the matrix as row major instead and the following stride-check failed.

Now the check is extended to identify both layouts. The only case where it cannot be exactly determined is a 1x1 matrix in which the following code just picks column major as default.
CC @cjnolet @tfeher

Authors:
  - Malte Förster (https://github.com/mfoerste4)

Approvers:
  - Corey J. Nolet (https://github.com/cjnolet)

URL: https://github.com/rapidsai/raft/pull/1521