Fixing ann bench cmake (and docs) (#1630)
I'm removing the build for the raft bfknn ann benchmark target for now since it doesn't seem to be complete anyways. I'm going to create a separate github issue to add it back it. For now, the cmake isn't even producing what the docs are saying it's producing so we need to fix that ASAP.

Authors:
  - Corey J. Nolet (https://github.com/cjnolet)

Approvers:
  - Dante Gama Dessavre (https://github.com/dantegd)

URL: https://github.com/rapidsai/raft/pull/1630