Compile libraft.so with separable compilation
With separable compilation, device code is linked together at the end of
compilation. This could be advantageous for RAFT, as we have many
duplicate kernel templates.

Previously, these duplicate kernel templates would be deduplicated on
the host side during linking, but not on device side, causing sometimes
up to 80 kernels to be left in the binary.

Now, both host and device side weak symbols (e.g., kernel templates)
should be deduplicated. Hopefully this has an impact on binary size.