Fixing Googletests and re-enabling in CI (#1904)
Recently, the CI script to run the C++ gtests was changed to call `ctest` but it looks like this command has been failing in CI for quite some time. It's a little scary as most of RAFT's C++ APIs are not exposed through Python and so we cannot rely on those Python tests to guarantee quality of the C++ APIs. 

I'm reverting the change back to invoking the C++ binaries directly so that we can get the tests running again. Once we figure out why `ctest` isn't running the tests then we can make the change to use that again.

Authors:
  - Corey J. Nolet (https://github.com/cjnolet)

Approvers:
  - William Hicks (https://github.com/wphicks)
  - Robert Maynard (https://github.com/robertmaynard)
  - Ray Douglass (https://github.com/raydouglass)
  - Ben Frederickson (https://github.com/benfred)

URL: https://github.com/rapidsai/raft/pull/1904