Reduce NN Descent test threshold (#1946)
Recently we have been noticing intermittent nightly failures in NN Descent tests that appear even when increasing the number of iterations. This is believed to be a result of non-determinism in the algorithm itself rather than a result of stochasticity, and so it cannot be seeded out. This PR attempts to fix the intermittent failures by slightly reducing the threshold.

Authors:
  - Divye Gala (https://github.com/divyegala)

Approvers:
  - Corey J. Nolet (https://github.com/cjnolet)

URL: https://github.com/rapidsai/raft/pull/1946