Add support for iterating over batches in bfknn
This adds support for iterating over batches of nearest neighbors in
the brute force knn. This lets you query for the nearest neighbors,
and then filter down the results - and if you have filtered out too many
results, get the next batch of nearest neighbors.

The challenge here is balancing memory consumption versus efficieny:
we could store the results of the full gemm in the distance calculation
- but for large indices this discards the benefits of using the tiling
strategy and risks running OOM. Instead we exponentially grow the number
of neighbors being returned, and also cache both the query norms and index
norms between calls.