Update select-k heuristic
With https://github.com/rapidsai/raft/pull/1878 merged, the performance
of the radix select algorithms is much improved and we no longer need
to incorporate the faiss block select algorithm. With #1878 merged,
faiss block select goes from being the 3rd ranked selection algorithm,
to the 5th.

This regenerates the heuristic function with the latest benchmark times,
and removes the faiss block select in favour of kWarpImmediate and
kRadix11bitsExtraPass.