Remove extraneous host pinnings from libraft-headers-only. (#2102)
We have `run` pinnings for librmm, spdlog, and fmt, so we don't need to list them in `host` as well. Currently, we are allowing `fmt>=10.1.1,<11` in `host`, which is pulling a newer version 10.2.1 and adding a run-export that is `fmt >=10.2.1,<11.0a0`, which disallows `fmt 10.1.1` contrary to our intentions.

Authors:
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - https://github.com/jakirkham
  - Ray Douglass (https://github.com/raydouglass)

URL: https://github.com/rapidsai/raft/pull/2102