Change the benchmark neighbors time size_t -> uint32_t and make benchmark loop event sync optional
When using the persistent kernel variant, the calling CPU thread has to synchronize
with the GPU (wait on the completion flag) - i.e. there's no way to use events for this.
As a result, the event recording and sync in the benchmark loop introduce significant latency overheads.
To avoid this, I make the event optional (dependant on the search mode: persistent/original).

Originally, the benchmark used size_t indices, whereas CAGRA operated with uint32_t.
As a result, we had to do a linear mapping (on GPU), which adds a kernel to the benchmark loop,
which goes against the event optimization above.
Hence, I changed the benchmark index type.