Set the import mode for dask tests (#2142)
This PR fixes the underlying issue that led us to pin pytest in #2137. For now, this PR leaves pytest pinned since there are other issues with pytest 8 and plugin interactions that have broken other RAPIDS libraries so I figured this was safer and we could relax the pin when the rest of RAPIDS does. However, if preferred I'm happy to undo the pinning in this PR too.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Corey J. Nolet (https://github.com/cjnolet)
  - Ray Douglass (https://github.com/raydouglass)

URL: https://github.com/rapidsai/raft/pull/2142