Drop NumPy from RAFT-Dask
The RAFT-Dask package doesn't actually `import numpy` anywhere inside.
This may have been different in the past. However today RAFT-Dask does
not use NumPy.

Further pylibraft, which does use NumPy and has a dependency on it, is
already a dependency of RAFT-Dask. As a result, NumPy is pulled into the
environment anyways.

The conda package for raft-dask already doesn't have NumPy and has been
working ok for a while. So line up the wheels to match this behavior.