Correct computation of Laplacian in spectral partitioning
In the previous implementation of spectral partitioning, the eigenvalue computation was actually performed on the original adjacency matrix, not the graph Laplacian thereof. This change provides a utility for computing the graph Laplacian, which is then fed into the new implementation of the Lanczos solver. An additional test is provided to ensure that the end-to-end spectral partitioning quality is as expected.