Fix static initialization order fiasco in `lanczos.cu`
`rows<>`/`cols<>`/`vals<>` are variable templates that are not
explicitly specialized. As such, for their initialization,
the rule that applies is "unordered dynamic initialization".
This means that their initialization is "indeterminately
sequenced with respect to all other dynamic initialization".

In particular, it can happen that `rows<int>` is initialized after its
use to initialize `inputsd`, which makes `lanczos_inputs.rows` empty,
with breaks tests (`rows.size()-1` underflows and results in `n==-1`,
leading to the test trying to allocate a large amount of memory.
Same for `cols<>` and `vals<>`.

This situation actually happens when building with clang. Fix it by
turning `rows`/`cols`/`vals` into a template function.