Fix laplacian scaling coefficients (#2871)
Resolves https://github.com/rapidsai/cuml/issues/7488

In a normalized laplacian scaling coefficients are the sqrt of the laplacian diagonal.
These coefficients are then used to scale all elements of the laplacian which ensure diagonal elements are set to 1 and off diagonal elements are scaled appropriately.
`(1/scaling)*L*(1/scaling)`

However, when the scaling coefficients are 0, it will cause a divide by 0 issue as seen above. This caused `nan` values to appear in the umap embedding output in cuml. The way to handle this is to substitute any 0 scaling values with 1 instead which ensures scaling will have no effect.

For more details see https://docs.scipy.org/doc/scipy/reference/generated/scipy.sparse.csgraph.laplacian.html and scroll to the `normed=True` section.

This PR adds a elementwise operation to replace the scaling zeros with ones.

Authors:
  - Anupam (https://github.com/aamijar)

Approvers:
  - Corey J. Nolet (https://github.com/cjnolet)
  - Divye Gala (https://github.com/divyegala)

URL: https://github.com/rapidsai/raft/pull/2871